fixed issues, code cleaning, renamed HeightsProvider into CellSizesProvider, fixed tests
